include rudimentary run script
